feat(client): add module utils
